requirements.txt
fastapi>=0.110.0
python-multipart>=0.0.6
uvicorn[standard]>=0.29.0
jinja2>=3.1.0
httpx>=0.27.0
pyyaml>=6.0
pydantic>=2.0.0
